+++ src/main/java/net/sf/mpxj/asta/AstaDatabaseFileReader.java
@@ -142,9 +142,11 @@ public final class AstaDatabaseFileReader extends AbstractProjectFileReader
    {
       List<Row> projectSummaryRows = getRows("select duration as durationhours, project_start as staru, project_end as ene, * from project_summary where projid=?", m_projectID);
       List<Row> progressPeriodRows = getRows("select id as progress_periodid, * from progress_period where projid=?", m_projectID);
+      List<Row> userSettingsRows = getRows("select * from userr where projid=?", m_projectID);      
       Row projectSummary = projectSummaryRows.isEmpty() ? null : projectSummaryRows.get(0);
+      Row userSettings = userSettingsRows.isEmpty() ? null : userSettingsRows.get(0);
       List<Row> progressPeriods = progressPeriodRows.isEmpty() ? null : progressPeriodRows;
-      m_reader.processProjectProperties(projectSummary, progressPeriods);
+      m_reader.processProjectProperties(projectSummary, userSettings, progressPeriods);
    }
 
    /**
@@ -494,6 +496,6 @@ public final class AstaDatabaseFileReader extends AbstractProjectFileReader
    private PreparedStatement m_ps;
    private ResultSet m_rs;
    private Map<String, Integer> m_meta = new HashMap<>();
-   
+
    private static final Integer DEFAULT_PROJECT_ID = Integer.valueOf(0);
 }