+++ src/main/java/net/sf/mpxj/asta/AstaDatabaseReader.java
@@ -132,9 +132,11 @@ public final class AstaDatabaseReader extends AbstractProjectFileReader
    {
       List<Row> projectSummaryRows = getRows("select * from project_summary where projid=?", m_projectID);
       List<Row> progressPeriodRows = getRows("select * from progress_period where projid=?", m_projectID);
+      List<Row> userSettingsRows = getRows("select * from userr where projid=?", m_projectID);      
       Row projectSummary = projectSummaryRows.isEmpty() ? null : projectSummaryRows.get(0);
+      Row userSettings = userSettingsRows.isEmpty() ? null : userSettingsRows.get(0);
       List<Row> progressPeriods = progressPeriodRows.isEmpty() ? null : progressPeriodRows;
-      m_reader.processProjectProperties(projectSummary, progressPeriods);
+      m_reader.processProjectProperties(projectSummary, userSettings, progressPeriods);
    }
 
    /**